last round of refactors
