+++ web/src/components/account/index.tsx
@@ -13,7 +13,7 @@ import {
 } from "@mui/material";
 import { useEffect, useState } from "react";
 import ExitToAppIcon from "@mui/icons-material/ExitToApp";
-import { AdminUser } from "../../model/adminUser";
+import { AdminUser } from "../../model/interfaces";
 import { useApi } from "../../contexts/apiContext";
 import { AxiosInstance } from "axios";
