+++ web/src/components/committee/components/adminControls/components/editDelegations/index.tsx
@@ -13,7 +13,7 @@ import {
   Typography,
 } from "@mui/material";
 import { useEffect, useState } from "react";
-import { Delegation } from "../../../../../../model/delegation";
+import { Delegation } from "../../../../../../model/interfaces";
 import { useApi } from "../../../../../../contexts/apiContext";
 import { useCommittee } from "../../../../contexts/committeeContext";
 import { useParams } from "react-router-dom";
