+++ web/src/components/delegation/index.tsx
@@ -1,7 +1,7 @@
 import { Box, CircularProgress } from "@mui/material";
 import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
-import { Delegation } from "../../model/delegation";
+import { Delegation } from "../../model/interfaces";
 import { useHeader } from "../../contexts/headerContext";
 import Widget from "../shared/widget";
 import CommitteesTable from "./components/committeesTable";
