--- web/src/model/adminUser.tsx
@@ -1,7 +1,0 @@
-export interface AdminUser {
-  user_id: number;
-  username: string;
-  first_name: string;
-  last_name: string;
-  super_user: boolean;
-}
