--- web/src/model/participant.tsx
@@ -1,1 +1,0 @@
-export interface Participant {}
