--- web/src/model/workingPaper.tsx
@@ -1,9 +1,0 @@
-import { Delegation } from "./delegation";
-
-export interface WorkingPaper {
-  working_paper_id: number;
-  paper_link: string;
-  committee_id: number;
-  working_group_name: string;
-  delegations: Delegation[];
-}
